backend/requirements.txt
fastapi
uvicorn
llama-parse
openai
python-dotenv
pydantic
pdfplumber
google-generativeai
